# PERF: avoid unneeded recoding of categoricals and reuse CategoricalDtypes for greater slicing speed