@topper-123 2018-06-27
The CircleCi error seems unrelated : ``AssertionError: Caused unexpected warning(s): ['ResourceWarning'].`` in pandas/tests/io/parser/dialect.py, which isn't part of this PR.